edited task routes with authentication validation

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 @Injectable()
 
 export class TasksService {
@@ -14,12 +15,15 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ){}
 
-    async getTasks(filterDto:GetTasksFilterDto):Promise<Task[]>{
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(
+        filterDto:GetTasksFilterDto,
+        user:User,
+        ):Promise<Task[]>{
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id:number):Promise<Task>{
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id:number, user:User):Promise<Task>{
+        const found = await this.taskRepository.findOne({where:{id, userId: user.id} });
 
         if(!found){
             // throw exception with custom error message when task with provided id is not found 
@@ -29,17 +33,20 @@ export class TasksService {
         return found
     }
 
-    async createTask(createTaskDto: CreateTaskDto):Promise<Task>{
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user:User,
+        ):Promise<Task>{
 
         //handled by repository of persistent data for database
-        return this.taskRepository.createTask(createTaskDto);
+        return this.taskRepository.createTask(createTaskDto, user);
     }
     
 
-    async deleteTask(id:number):Promise<void>{
+    async deleteTask(id:number, user:User):Promise<void>{
         //typeorm delete method passes element inside entity 
         // ie.id and delete and return raw or effected types (recomended for performance instead of remove method) 
-        const result = await this.taskRepository.delete(id);
+        const result = await this.taskRepository.delete({id, userId: user.id});
 
         if(result.affected === 0){
             // throw exception with custom error message when task with provided id is not found (affected is 0)
@@ -48,9 +55,9 @@ export class TasksService {
     }
 
     //setting aways setting return type as promise because its a syncronous method
-    async updateTaskStatus(id:number, status: TaskStatus):Promise<Task>{
+    async updateTaskStatus(id:number, status: TaskStatus,  user:User):Promise<Task>{
         // calling get by id function to get task first
-        const task = await this.getTaskById(id);
+        const task = await this.getTaskById(id, user);
         // updating status status 
         task.status = status;
         await task.save();
